# Add more methods to `quat`