--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -3,7 +3,9 @@ using Quaternions
 using Aqua
 using RealDot
 
-Aqua.test_all(Quaternions)
+# Avoid tests on unbound_args.
+# https://github.com/JuliaGeometry/Quaternions.jl/pull/132#discussion_r1383817950
+Aqua.test_all(Quaternions, unbound_args=false)
 
 include("helpers.jl")
 include("Quaternion.jl")